+++ crates/networking/p2p/sync.rs
@@ -1,12 +1,15 @@
+mod code_collector;
 mod state_healing;
 mod storage_healing;
 
 use crate::peer_handler::{BlockRequestOrder, PeerHandlerError, SNAP_LIMIT};
 use crate::rlpx::p2p::SUPPORTED_ETH_CAPABILITIES;
+use crate::sync::code_collector::CodeHashCollector;
 use crate::sync::state_healing::heal_state_trie_wrap;
 use crate::sync::storage_healing::heal_storage_trie;
 use crate::utils::{
     current_unix_time, get_account_state_snapshots_dir, get_account_storages_snapshots_dir,
+    get_code_hashes_snapshots_dir,
 };
 use crate::{
     metrics::METRICS,
@@ -846,6 +849,12 @@ impl Syncer {
         let account_state_snapshots_dir = get_account_state_snapshots_dir(&self.datadir);
         let account_storages_snapshots_dir = get_account_storages_snapshots_dir(&self.datadir);
 
+        let code_hashes_snapshot_dir = get_code_hashes_snapshots_dir(&self.datadir);
+        std::fs::create_dir_all(&code_hashes_snapshot_dir).map_err(|_| SyncError::CorruptPath)?;
+
+        // Create collector to store code hashes in files
+        let mut code_hash_collector = CodeHashCollector::new(code_hashes_snapshot_dir.clone());
+
         let mut storage_accounts = AccountStorageRoots::default();
         if !std::env::var("SKIP_START_SNAP_SYNC").is_ok_and(|var| !var.is_empty()) {
             // We start by downloading all of the leafs of the trie of accounts
@@ -898,6 +907,17 @@ impl Syncer {
                 );
 
                 info!("Inserting accounts into the state trie");
+
+                // Collect valid code hashes from current account snapshot
+                let code_hashes_from_snapshot: Vec<H256> = account_states_snapshot
+                    .iter()
+                    .filter_map(|(_, state)| {
+                        (state.code_hash != *EMPTY_KECCACK_HASH).then_some(state.code_hash)
+                    })
+                    .collect();
+
+                code_hash_collector.extend(code_hashes_from_snapshot);
+                code_hash_collector.flush_if_needed().await?;
 
                 let store_clone = store.clone();
                 let current_state_root =
@@ -957,6 +977,7 @@ impl Syncer {
                     calculate_staleness_timestamp(pivot_header.timestamp),
                     &mut state_leafs_healed,
                     &mut storage_accounts,
+                    &mut code_hash_collector,
                 )
                 .await?
                 {
@@ -1082,6 +1103,7 @@ impl Syncer {
                 calculate_staleness_timestamp(pivot_header.timestamp),
                 &mut global_state_leafs_healed,
                 &mut storage_accounts,
+                &mut code_hash_collector,
             )
             .await?;
             if !healing_done {
@@ -1106,31 +1128,70 @@ impl Syncer {
         debug_assert!(validate_storage_root(store.clone(), pivot_header.state_root).await);
         info!("Finished healing");
 
+        // Finish code hash collection
+        code_hash_collector.finish().await?;
+
         *METRICS.bytecode_download_start_time.lock().await = Some(SystemTime::now());
-        let mut bytecode_iter = store
-            .iter_accounts(pivot_header.state_root)
-            .expect("we couldn't iterate over accounts")
-            .map(|(_, state)| state.code_hash)
-            .filter(|code_hash| *code_hash != *EMPTY_KECCACK_HASH);
-        for mut bytecode_hashes in std::iter::from_fn(|| bytecode_iter_fn(&mut bytecode_iter)) {
-            // Download bytecodes
-            bytecode_hashes.sort();
-            bytecode_hashes.dedup();
-            info!(
-                "Starting bytecode download of {} hashes",
-                bytecode_hashes.len()
-            );
+
+        let code_hashes_dir = get_code_hashes_snapshots_dir(&self.datadir);
+        let mut seen_code_hashes = HashSet::new();
+        let mut code_hashes_to_download = Vec::new();
+
+        info!("Starting download code hashes from peers");
+        for entry in std::fs::read_dir(&code_hashes_dir)
+            .map_err(|_| SyncError::CodeHashesSnapshotsDirNotFound)?
+        {
+            let entry = entry.map_err(|_| SyncError::CorruptPath)?;
+            let snapshot_contents = std::fs::read(entry.path())
+                .map_err(|err| SyncError::SnapshotReadError(entry.path(), err))?;
+            let code_hashes: Vec<H256> = RLPDecode::decode(&snapshot_contents)
+                .map_err(|_| SyncError::CodeHashesSnapshotDecodeError(entry.path()))?;
+
+            for hash in code_hashes {
+                // If we haven't seen the code hash yet, add it to the list of hashes to download
+                if seen_code_hashes.insert(hash) {
+                    code_hashes_to_download.push(hash);
+
+                    if code_hashes_to_download.len() >= BYTECODE_CHUNK_SIZE {
+                        info!(
+                            "Starting bytecode download of {} hashes",
+                            code_hashes_to_download.len()
+                        );
+                        let bytecodes = self
+                            .peers
+                            .request_bytecodes(&code_hashes_to_download)
+                            .await
+                            .map_err(SyncError::PeerHandler)?
+                            .ok_or(SyncError::BytecodesNotFound)?;
+
+                        store
+                            .write_account_code_batch(
+                                code_hashes_to_download.drain(..).zip(bytecodes).collect(),
+                            )
+                            .await?;
+                    }
+                }
+            }
+        }
+
+        // Download remaining bytecodes if any
+        if !code_hashes_to_download.is_empty() {
             let bytecodes = self
                 .peers
-                .request_bytecodes(&bytecode_hashes)
+                .request_bytecodes(&code_hashes_to_download)
                 .await
                 .map_err(SyncError::PeerHandler)?
                 .ok_or(SyncError::BytecodesNotFound)?;
             store
-                .write_account_code_batch(bytecode_hashes.into_iter().zip(bytecodes).collect())
+                .write_account_code_batch(
+                    code_hashes_to_download.into_iter().zip(bytecodes).collect(),
+                )
                 .await?;
         }
+
         *METRICS.bytecode_download_end_time.lock().await = Some(SystemTime::now());
+
+        debug_assert!(validate_bytecodes(store.clone(), pivot_header.state_root).await);
 
         store_block_bodies(vec![pivot_header.hash()], self.peers.clone(), store.clone()).await?;
 
@@ -1313,6 +1374,8 @@ pub enum SyncError {
     SnapshotReadError(PathBuf, std::io::Error),
     #[error("Failed to RLP decode account_state_snapshot from {0:?}")]
     SnapshotDecodeError(PathBuf),
+    #[error("Failed to RLP decode code_hashes_snapshot from {0:?}")]
+    CodeHashesSnapshotDecodeError(PathBuf),
     #[error("Failed to get account state for block {0:?} and account hash {1:?}")]
     AccountState(H256, H256),
     #[error("Failed to acquire lock on maybe_big_account_storage")]
@@ -1323,6 +1386,8 @@ pub enum SyncError {
     AccountStateSnapshotsDirNotFound,
     #[error("Failed to get account storages snapshots directory")]
     AccountStoragesSnapshotsDirNotFound,
+    #[error("Failed to get code hashes snapshots directory")]
+    CodeHashesSnapshotsDirNotFound,
     #[error("Got different state roots for account hash: {0:?}, expected: {1:?}, computed: {2:?}")]
     DifferentStateRoots(H256, H256, H256),
     #[error("We aren't finding get_peer_channel_with_retry")]
@@ -1335,6 +1400,8 @@ pub enum SyncError {
     PeerHandler(#[from] PeerHandlerError),
     #[error("Corrupt Path")]
     CorruptPath,
+    #[error("Bytecode file error")]
+    BytecodeFileError,
 }
 
 impl<T> From<SendError<T>> for SyncError {
@@ -1403,10 +1470,27 @@ pub async fn validate_storage_root(store: Store, state_root: H256) -> bool {
     is_valid
 }
 
-fn bytecode_iter_fn<T>(bytecode_iter: &mut T) -> Option<Vec<H256>>
-where
-    T: Iterator<Item = H256>,
-{
-    Some(bytecode_iter.by_ref().take(BYTECODE_CHUNK_SIZE).collect())
-        .filter(|chunk: &Vec<_>| !chunk.is_empty())
+pub async fn validate_bytecodes(store: Store, state_root: H256) -> bool {
+    info!("Starting validate_bytecodes");
+    let mut is_valid = true;
+    for (account_hash, account_state) in store
+        .iter_accounts(state_root)
+        .expect("we couldn't iterate over accounts")
+    {
+        if account_state.code_hash != *EMPTY_KECCACK_HASH
+            && !store
+                .get_account_code(account_state.code_hash)
+                .is_ok_and(|code| code.is_some())
+        {
+            error!(
+                "Missing code hash {:x} for account {:x}",
+                account_state.code_hash, account_hash
+            );
+            is_valid = false
+        }
+    }
+    if !is_valid {
+        std::process::exit(-1);
+    }
+    is_valid
 }